New Version of Collected Photo UI
In this version when user collects a photo with search option, user can
stay at the same search page but also have a button which is linked to
photo collection page. This version fixes the bug of nonfunctional edit
and delete buttons after photo collected, with a brand new button. Also
it gives a better user experience bec.of the button which is linked to
user photo collection page.